gif_certificate/service/src/test/java/com/epam/esm/validator/TagValidatorTest.java
package com.epam.esm.validator;

import com.epam.esm.entity.Tag;
import com.epam.esm.exceptions.IncorrectParameterException;
import com.epam.esm.validator.TagValidator;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
import static org.junit.jupiter.api.Assertions.assertThrows;

public class TagValidatorTest {
    private static final String INCORRECT_NAME = "qw";
    private static final String CORRECT_NAME = "tagName";
    private static final Tag INCORRECT_TAG = new Tag(1L, INCORRECT_NAME);
    private static final Tag CORRECT_TAG = new Tag(2L, CORRECT_NAME);

    @Test
    void testValidate_incorrectData() {
        assertThrows(IncorrectParameterException.class, () -> TagValidator.validate(INCORRECT_TAG));
    }

    @Test
    void testValidate_correctData() {
        assertDoesNotThrow(() -> TagValidator.validate(CORRECT_TAG));
    }

    @Test
    void testValidateName_incorrectData() {
        assertThrows(IncorrectParameterException.class, () -> TagValidator.validateName(INCORRECT_NAME));
    }

    @Test
    void testValidateName_correctData() {
        assertDoesNotThrow(() -> TagValidator.validateName(CORRECT_NAME));
    }
}
